\name{SwissInhabitants}
\alias{SwissInhabitants}
\alias{SwissInhabitants1900}

\docType{data}
\title{
SwissInhabitants in 1900
}
\description{
Number of inhabitants in the 2896 villages of Switzerland in the year 1900. 
}
\usage{data("SwissInhabitants")}

\details{
This data set consists of the number of inhabitants in the 2896 communes, i.e. cities and villages, in the year 1900.
The individual count is the total number of persons living in the particular commune. The data set is unordered for
anonymity reasons. The data set has been used as part of a larger data set to identify patterns of concentration in Switzerland (see reference).
}
\source{
Schuler,M.,  Ullmann, D. Eidgenossische Volkszahlung:Bevoelkerungsentwicklung der Gemeinden, Bundesamt fur Statistik, Neuchatel, Switzerland, 2002 
}
\references{
Behnisch, M., Ultsch, A.: Population Patterns in Switzerland 1850-2000, in: Gaul, W. et al (Eds), Advances in Data Analysis, Data Handling and Business Intelligence, Springer, Heidelberg, pp. 163-173, 2010.
}
\examples{
data(SwissInhabitants)
## maybe str(SwissInhabitants) ; plot(SwissInhabitants) ...
}
\keyword{datasets,SwissInhabitants,SwissInhabitants1900}
